LPS-184431 Move packages
